Refactoring of Modal, added success modal, styling
