+++ styles/chatBox.module.css
@@ -192,6 +192,7 @@
   width: 100%;
   color: lightgray;
   margin-right: 3%;
+  cursor: default;
 }
 
 .likeIcon {
@@ -214,4 +215,33 @@
   height: 100%;
   width: 100%;
   color: lightgray;
+  cursor: default;
+}
+
+.feedback {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.feedbackBox {
+  margin: 1rem;
+  border: none;
+}
+
+.input {
+  font-size: 15px;
+  font-family: "Open Sans", cursive;
+}
+
+.submit {
+  width: 17%;
+  padding: 0.5%;
+  background-color: #77a465;
+  border: none;
+  border-radius: 7px;
+  cursor: pointer;
+  font-size: 15px;
+  color: white;
+  font-family: "Open Sans", cursive;
 }
