+++ styles/modal.module.css
@@ -47,34 +47,6 @@
   align-items: center;
 }
 
-.feedback {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.feedbackBox {
-  margin: 1rem;
-  border: none;
-}
-
-.input {
-  font-size: 15px;
-  font-family: "Open Sans", cursive;
-}
-
-.button {
-  width: 17%;
-  padding: 0.5%;
-  background-color: #77a465;
-  border: none;
-  border-radius: 7px;
-  cursor: pointer;
-  font-size: 15px;
-  color: white;
-  font-family: "Open Sans", cursive;
-}
-
 @media (min-width: 768px) {
   .modal {
     width: 40rem;
